chore: cleanup and fix integration tests
