--- db/schema.ts
@@ -1,1 +1,0 @@
-export * from './drizzle/schema';
