+++ integration/tests/integration.test.ts
@@ -378,7 +378,9 @@ describe('types', () => {
     expect(result?.enumArray).toStrictEqual(['pending', 'active']);
 
     const dbResult = await db.query.allTypes.findFirst({
-      where: (table, {eq}) => eq(table.id, '1011'),
+      where: {
+        id: '1011',
+      },
     });
 
     expect(dbResult?.id).toStrictEqual('1011');
