+++ integration/zero-schema.gen.ts
@@ -1,4 +1,4 @@
-// @generated drizzle-zero signature:sha256:ecf7cbf48d18475e3638cda4ed15e81cb3157f5b744eb02d705f16a9de1f0a60
+// @generated drizzle-zero signature:sha256:04b41f784215311f7b546b49c712771c1d28e1139d68cd92110f2f42e1938da4
 // This file was automatically generated by drizzle-zero.
 // You should NOT make any changes in this file as it will be overwritten.
 
@@ -7410,101 +7410,127 @@ const shipmentItemRelationships = {
     },
   ],
 } as const;
-const userRelationships = {
-  messages: [
+const benefitEnrollmentRelationships = {
+  benefitPlan: [
     {
-      sourceField: ['id'],
-      destField: ['senderId'],
-      destSchema: 'message',
-      cardinality: 'many',
-    },
-  ],
-  mediums: [
-    {
-      sourceField: ['id'],
-      destField: ['senderId'],
-      destSchema: 'message',
-      cardinality: 'many',
-    },
-    {
-      sourceField: ['mediumId'],
+      sourceField: ['benefitPlanId'],
       destField: ['id'],
-      destSchema: 'medium',
-      cardinality: 'many',
-    },
-  ],
-  friends: [
-    {
-      sourceField: ['id'],
-      destField: ['requestingId'],
-      destSchema: 'friendship',
-      cardinality: 'many',
-    },
-    {
-      sourceField: ['acceptingId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const mediumRelationships = {
-  messages: [
-    {
-      sourceField: ['id'],
-      destField: ['mediumId'],
-      destSchema: 'message',
-      cardinality: 'many',
-    },
-  ],
-  users: [
-    {
-      sourceField: ['id'],
-      destField: ['mediumId'],
-      destSchema: 'message',
-      cardinality: 'many',
-    },
-    {
-      sourceField: ['senderId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const messageRelationships = {
-  medium: [
-    {
-      sourceField: ['mediumId'],
-      destField: ['id'],
-      destSchema: 'medium',
+      destSchema: 'benefitPlan',
       cardinality: 'one',
     },
   ],
-  sender: [
+  employee: [
     {
-      sourceField: ['senderId'],
+      sourceField: ['employeeId'],
+      destField: ['id'],
+      destSchema: 'employeeProfile',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const benefitPlanRelationships = {
+  administrator: [
+    {
+      sourceField: ['administratorId'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
     },
   ],
-} as const;
-const filtersRelationships = {
-  parent: [
+  enrollments: [
     {
-      sourceField: ['parentId'],
+      sourceField: ['id'],
+      destField: ['benefitPlanId'],
+      destSchema: 'benefitEnrollment',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const billingInvoiceRelationships = {
+  account: [
+    {
+      sourceField: ['accountId'],
       destField: ['id'],
-      destSchema: 'filters',
+      destSchema: 'crmAccount',
       cardinality: 'one',
     },
   ],
-  children: [
+  contact: [
+    {
+      sourceField: ['contactId'],
+      destField: ['id'],
+      destSchema: 'crmContact',
+      cardinality: 'one',
+    },
+  ],
+  issuer: [
+    {
+      sourceField: ['issuedById'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+  lines: [
     {
       sourceField: ['id'],
-      destField: ['parentId'],
-      destSchema: 'filters',
+      destField: ['invoiceId'],
+      destSchema: 'billingInvoiceLine',
       cardinality: 'many',
+    },
+  ],
+} as const;
+const billingInvoiceLineRelationships = {
+  invoice: [
+    {
+      sourceField: ['invoiceId'],
+      destField: ['id'],
+      destSchema: 'billingInvoice',
+      cardinality: 'one',
+    },
+  ],
+  orderItem: [
+    {
+      sourceField: ['orderItemId'],
+      destField: ['id'],
+      destSchema: 'orderItem',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const budgetRelationships = {
+  department: [
+    {
+      sourceField: ['departmentId'],
+      destField: ['id'],
+      destSchema: 'department',
+      cardinality: 'one',
+    },
+  ],
+  lines: [
+    {
+      sourceField: ['id'],
+      destField: ['budgetId'],
+      destSchema: 'budgetLine',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const budgetLineRelationships = {
+  budget: [
+    {
+      sourceField: ['budgetId'],
+      destField: ['id'],
+      destSchema: 'budget',
+      cardinality: 'one',
+    },
+  ],
+  account: [
+    {
+      sourceField: ['accountId'],
+      destField: ['id'],
+      destSchema: 'ledgerAccount',
+      cardinality: 'one',
     },
   ],
 } as const;
@@ -7550,6 +7576,58 @@ const crmAccountRelationships = {
     },
   ],
 } as const;
+const crmActivityRelationships = {
+  account: [
+    {
+      sourceField: ['accountId'],
+      destField: ['id'],
+      destSchema: 'crmAccount',
+      cardinality: 'one',
+    },
+  ],
+  contact: [
+    {
+      sourceField: ['contactId'],
+      destField: ['id'],
+      destSchema: 'crmContact',
+      cardinality: 'one',
+    },
+  ],
+  opportunity: [
+    {
+      sourceField: ['opportunityId'],
+      destField: ['id'],
+      destSchema: 'crmOpportunity',
+      cardinality: 'one',
+    },
+  ],
+  type: [
+    {
+      sourceField: ['typeId'],
+      destField: ['id'],
+      destSchema: 'crmActivityType',
+      cardinality: 'one',
+    },
+  ],
+  performer: [
+    {
+      sourceField: ['performedById'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const crmActivityTypeRelationships = {
+  activities: [
+    {
+      sourceField: ['id'],
+      destField: ['typeId'],
+      destSchema: 'crmActivity',
+      cardinality: 'many',
+    },
+  ],
+} as const;
 const crmContactRelationships = {
   account: [
     {
@@ -7576,21 +7654,29 @@ const crmContactRelationships = {
     },
   ],
 } as const;
-const crmPipelineStageRelationships = {
-  opportunities: [
+const crmNoteRelationships = {
+  account: [
     {
-      sourceField: ['id'],
-      destField: ['stageId'],
-      destSchema: 'crmOpportunity',
-      cardinality: 'many',
+      sourceField: ['accountId'],
+      destField: ['id'],
+      destSchema: 'crmAccount',
+      cardinality: 'one',
     },
   ],
-  historyEntries: [
+  contact: [
     {
-      sourceField: ['id'],
-      destField: ['stageId'],
-      destSchema: 'crmOpportunityStageHistory',
-      cardinality: 'many',
+      sourceField: ['contactId'],
+      destField: ['id'],
+      destSchema: 'crmContact',
+      cardinality: 'one',
+    },
+  ],
+  author: [
+    {
+      sourceField: ['authorId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
     },
   ],
 } as const;
@@ -7654,132 +7740,46 @@ const crmOpportunityStageHistoryRelationships = {
     },
   ],
 } as const;
-const crmActivityTypeRelationships = {
-  activities: [
+const crmPipelineStageRelationships = {
+  opportunities: [
     {
       sourceField: ['id'],
-      destField: ['typeId'],
-      destSchema: 'crmActivity',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const crmActivityRelationships = {
-  account: [
-    {
-      sourceField: ['accountId'],
-      destField: ['id'],
-      destSchema: 'crmAccount',
-      cardinality: 'one',
-    },
-  ],
-  contact: [
-    {
-      sourceField: ['contactId'],
-      destField: ['id'],
-      destSchema: 'crmContact',
-      cardinality: 'one',
-    },
-  ],
-  opportunity: [
-    {
-      sourceField: ['opportunityId'],
-      destField: ['id'],
+      destField: ['stageId'],
       destSchema: 'crmOpportunity',
-      cardinality: 'one',
+      cardinality: 'many',
     },
   ],
-  type: [
+  historyEntries: [
     {
-      sourceField: ['typeId'],
-      destField: ['id'],
-      destSchema: 'crmActivityType',
-      cardinality: 'one',
+      sourceField: ['id'],
+      destField: ['stageId'],
+      destSchema: 'crmOpportunityStageHistory',
+      cardinality: 'many',
     },
   ],
-  performer: [
+} as const;
+const departmentRelationships = {
+  manager: [
     {
-      sourceField: ['performedById'],
+      sourceField: ['managerId'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
     },
   ],
-} as const;
-const crmNoteRelationships = {
-  account: [
-    {
-      sourceField: ['accountId'],
-      destField: ['id'],
-      destSchema: 'crmAccount',
-      cardinality: 'one',
-    },
-  ],
-  contact: [
-    {
-      sourceField: ['contactId'],
-      destField: ['id'],
-      destSchema: 'crmContact',
-      cardinality: 'one',
-    },
-  ],
-  author: [
-    {
-      sourceField: ['authorId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const documentLibraryRelationships = {
-  project: [
-    {
-      sourceField: ['projectId'],
-      destField: ['id'],
-      destSchema: 'project',
-      cardinality: 'one',
-    },
-  ],
-  folders: [
+  teams: [
     {
       sourceField: ['id'],
-      destField: ['libraryId'],
-      destSchema: 'documentFolder',
+      destField: ['departmentId'],
+      destSchema: 'team',
       cardinality: 'many',
     },
   ],
-} as const;
-const documentFolderRelationships = {
-  library: [
-    {
-      sourceField: ['libraryId'],
-      destField: ['id'],
-      destSchema: 'documentLibrary',
-      cardinality: 'one',
-    },
-  ],
-  parent: [
-    {
-      sourceField: ['parentId'],
-      destField: ['id'],
-      destSchema: 'documentFolder',
-      cardinality: 'one',
-    },
-  ],
-  children: [
+  employees: [
     {
       sourceField: ['id'],
-      destField: ['parentId'],
-      destSchema: 'documentFolder',
-      cardinality: 'many',
-    },
-  ],
-  files: [
-    {
-      sourceField: ['id'],
-      destField: ['folderId'],
-      destSchema: 'documentFile',
+      destField: ['departmentId'],
+      destSchema: 'employeeProfile',
       cardinality: 'many',
     },
   ],
@@ -7836,6 +7836,58 @@ const documentFileVersionRelationships = {
     },
   ],
 } as const;
+const documentFolderRelationships = {
+  library: [
+    {
+      sourceField: ['libraryId'],
+      destField: ['id'],
+      destSchema: 'documentLibrary',
+      cardinality: 'one',
+    },
+  ],
+  parent: [
+    {
+      sourceField: ['parentId'],
+      destField: ['id'],
+      destSchema: 'documentFolder',
+      cardinality: 'one',
+    },
+  ],
+  children: [
+    {
+      sourceField: ['id'],
+      destField: ['parentId'],
+      destSchema: 'documentFolder',
+      cardinality: 'many',
+    },
+  ],
+  files: [
+    {
+      sourceField: ['id'],
+      destField: ['folderId'],
+      destSchema: 'documentFile',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const documentLibraryRelationships = {
+  project: [
+    {
+      sourceField: ['projectId'],
+      destField: ['id'],
+      destSchema: 'project',
+      cardinality: 'one',
+    },
+  ],
+  folders: [
+    {
+      sourceField: ['id'],
+      destField: ['libraryId'],
+      destSchema: 'documentFolder',
+      cardinality: 'many',
+    },
+  ],
+} as const;
 const documentSharingRelationships = {
   file: [
     {
@@ -7862,351 +7914,21 @@ const documentSharingRelationships = {
     },
   ],
 } as const;
-const billingInvoiceRelationships = {
-  account: [
+const employeeDocumentRelationships = {
+  employee: [
     {
-      sourceField: ['accountId'],
+      sourceField: ['employeeId'],
       destField: ['id'],
-      destSchema: 'crmAccount',
-      cardinality: 'one',
-    },
-  ],
-  contact: [
-    {
-      sourceField: ['contactId'],
-      destField: ['id'],
-      destSchema: 'crmContact',
-      cardinality: 'one',
-    },
-  ],
-  issuer: [
-    {
-      sourceField: ['issuedById'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  lines: [
-    {
-      sourceField: ['id'],
-      destField: ['invoiceId'],
-      destSchema: 'billingInvoiceLine',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const billingInvoiceLineRelationships = {
-  invoice: [
-    {
-      sourceField: ['invoiceId'],
-      destField: ['id'],
-      destSchema: 'billingInvoice',
-      cardinality: 'one',
-    },
-  ],
-  orderItem: [
-    {
-      sourceField: ['orderItemId'],
-      destField: ['id'],
-      destSchema: 'orderItem',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const expenseReportRelationships = {
-  owner: [
-    {
-      sourceField: ['ownerId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  department: [
-    {
-      sourceField: ['departmentId'],
-      destField: ['id'],
-      destSchema: 'department',
-      cardinality: 'one',
-    },
-  ],
-  items: [
-    {
-      sourceField: ['id'],
-      destField: ['reportId'],
-      destSchema: 'expenseItem',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const expenseItemRelationships = {
-  report: [
-    {
-      sourceField: ['reportId'],
-      destField: ['id'],
-      destSchema: 'expenseReport',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const ledgerAccountRelationships = {
-  parent: [
-    {
-      sourceField: ['parentAccountId'],
-      destField: ['id'],
-      destSchema: 'ledgerAccount',
-      cardinality: 'one',
-    },
-  ],
-  children: [
-    {
-      sourceField: ['id'],
-      destField: ['parentAccountId'],
-      destSchema: 'ledgerAccount',
-      cardinality: 'many',
-    },
-  ],
-  entries: [
-    {
-      sourceField: ['id'],
-      destField: ['accountId'],
-      destSchema: 'ledgerEntry',
-      cardinality: 'many',
-    },
-  ],
-  budgetLines: [
-    {
-      sourceField: ['id'],
-      destField: ['accountId'],
-      destSchema: 'budgetLine',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const ledgerTransactionRelationships = {
-  creator: [
-    {
-      sourceField: ['createdById'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  entries: [
-    {
-      sourceField: ['id'],
-      destField: ['transactionId'],
-      destSchema: 'ledgerEntry',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const ledgerEntryRelationships = {
-  transaction: [
-    {
-      sourceField: ['transactionId'],
-      destField: ['id'],
-      destSchema: 'ledgerTransaction',
-      cardinality: 'one',
-    },
-  ],
-  account: [
-    {
-      sourceField: ['accountId'],
-      destField: ['id'],
-      destSchema: 'ledgerAccount',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const budgetRelationships = {
-  department: [
-    {
-      sourceField: ['departmentId'],
-      destField: ['id'],
-      destSchema: 'department',
-      cardinality: 'one',
-    },
-  ],
-  lines: [
-    {
-      sourceField: ['id'],
-      destField: ['budgetId'],
-      destSchema: 'budgetLine',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const budgetLineRelationships = {
-  budget: [
-    {
-      sourceField: ['budgetId'],
-      destField: ['id'],
-      destSchema: 'budget',
-      cardinality: 'one',
-    },
-  ],
-  account: [
-    {
-      sourceField: ['accountId'],
-      destField: ['id'],
-      destSchema: 'ledgerAccount',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const integrationWebhookRelationships = {
-  project: [
-    {
-      sourceField: ['projectId'],
-      destField: ['id'],
-      destSchema: 'project',
-      cardinality: 'one',
-    },
-  ],
-  account: [
-    {
-      sourceField: ['accountId'],
-      destField: ['id'],
-      destSchema: 'crmAccount',
-      cardinality: 'one',
-    },
-  ],
-  events: [
-    {
-      sourceField: ['id'],
-      destField: ['webhookId'],
-      destSchema: 'integrationEvent',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const integrationEventRelationships = {
-  webhook: [
-    {
-      sourceField: ['webhookId'],
-      destField: ['id'],
-      destSchema: 'integrationWebhook',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const integrationCredentialRelationships = {
-  webhook: [
-    {
-      sourceField: ['webhookId'],
-      destField: ['id'],
-      destSchema: 'integrationWebhook',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const marketingCampaignRelationships = {
-  owner: [
-    {
-      sourceField: ['ownerId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  channels: [
-    {
-      sourceField: ['id'],
-      destField: ['campaignId'],
-      destSchema: 'marketingCampaignChannel',
-      cardinality: 'many',
-    },
-  ],
-  audiences: [
-    {
-      sourceField: ['id'],
-      destField: ['campaignId'],
-      destSchema: 'marketingCampaignAudience',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const marketingChannelRelationships = {
-  campaignChannels: [
-    {
-      sourceField: ['id'],
-      destField: ['channelId'],
-      destSchema: 'marketingCampaignChannel',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const marketingCampaignChannelRelationships = {
-  campaign: [
-    {
-      sourceField: ['campaignId'],
-      destField: ['id'],
-      destSchema: 'marketingCampaign',
-      cardinality: 'one',
-    },
-  ],
-  channel: [
-    {
-      sourceField: ['channelId'],
-      destField: ['id'],
-      destSchema: 'marketingChannel',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const marketingAudienceRelationships = {
-  campaignAudiences: [
-    {
-      sourceField: ['id'],
-      destField: ['audienceId'],
-      destSchema: 'marketingCampaignAudience',
-      cardinality: 'many',
-    },
-  ],
-} as const;
-const marketingCampaignAudienceRelationships = {
-  campaign: [
-    {
-      sourceField: ['campaignId'],
-      destField: ['id'],
-      destSchema: 'marketingCampaign',
-      cardinality: 'one',
-    },
-  ],
-  audience: [
-    {
-      sourceField: ['audienceId'],
-      destField: ['id'],
-      destSchema: 'marketingAudience',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const departmentRelationships = {
-  manager: [
-    {
-      sourceField: ['managerId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  teams: [
-    {
-      sourceField: ['id'],
-      destField: ['departmentId'],
-      destSchema: 'team',
-      cardinality: 'many',
-    },
-  ],
-  employees: [
-    {
-      sourceField: ['id'],
-      destField: ['departmentId'],
       destSchema: 'employeeProfile',
-      cardinality: 'many',
+      cardinality: 'one',
+    },
+  ],
+  uploader: [
+    {
+      sourceField: ['uploadedById'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
     },
   ],
 } as const;
@@ -8268,32 +7990,6 @@ const employeeProfileRelationships = {
     },
   ],
 } as const;
-const teamRelationships = {
-  department: [
-    {
-      sourceField: ['departmentId'],
-      destField: ['id'],
-      destSchema: 'department',
-      cardinality: 'one',
-    },
-  ],
-  lead: [
-    {
-      sourceField: ['leadId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-  employees: [
-    {
-      sourceField: ['id'],
-      destField: ['teamId'],
-      destSchema: 'employeeProfile',
-      cardinality: 'many',
-    },
-  ],
-} as const;
 const employmentHistoryRelationships = {
   employee: [
     {
@@ -8304,36 +8000,162 @@ const employmentHistoryRelationships = {
     },
   ],
 } as const;
-const employeeDocumentRelationships = {
-  employee: [
+const expenseItemRelationships = {
+  report: [
     {
-      sourceField: ['employeeId'],
+      sourceField: ['reportId'],
       destField: ['id'],
-      destSchema: 'employeeProfile',
+      destSchema: 'expenseReport',
       cardinality: 'one',
     },
   ],
-  uploader: [
+} as const;
+const expenseReportRelationships = {
+  owner: [
     {
-      sourceField: ['uploadedById'],
+      sourceField: ['ownerId'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
     },
   ],
-} as const;
-const timesheetRelationships = {
-  employee: [
+  department: [
     {
-      sourceField: ['employeeId'],
+      sourceField: ['departmentId'],
       destField: ['id'],
-      destSchema: 'employeeProfile',
+      destSchema: 'department',
       cardinality: 'one',
     },
   ],
-  submittedBy: [
+  items: [
     {
-      sourceField: ['submittedById'],
+      sourceField: ['id'],
+      destField: ['reportId'],
+      destSchema: 'expenseItem',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const filtersRelationships = {
+  parent: [
+    {
+      sourceField: ['parentId'],
+      destField: ['id'],
+      destSchema: 'filters',
+      cardinality: 'one',
+    },
+  ],
+  children: [
+    {
+      sourceField: ['id'],
+      destField: ['parentId'],
+      destSchema: 'filters',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const integrationCredentialRelationships = {
+  webhook: [
+    {
+      sourceField: ['webhookId'],
+      destField: ['id'],
+      destSchema: 'integrationWebhook',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const integrationEventRelationships = {
+  webhook: [
+    {
+      sourceField: ['webhookId'],
+      destField: ['id'],
+      destSchema: 'integrationWebhook',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const integrationWebhookRelationships = {
+  project: [
+    {
+      sourceField: ['projectId'],
+      destField: ['id'],
+      destSchema: 'project',
+      cardinality: 'one',
+    },
+  ],
+  account: [
+    {
+      sourceField: ['accountId'],
+      destField: ['id'],
+      destSchema: 'crmAccount',
+      cardinality: 'one',
+    },
+  ],
+  events: [
+    {
+      sourceField: ['id'],
+      destField: ['webhookId'],
+      destSchema: 'integrationEvent',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const ledgerAccountRelationships = {
+  parent: [
+    {
+      sourceField: ['parentAccountId'],
+      destField: ['id'],
+      destSchema: 'ledgerAccount',
+      cardinality: 'one',
+    },
+  ],
+  children: [
+    {
+      sourceField: ['id'],
+      destField: ['parentAccountId'],
+      destSchema: 'ledgerAccount',
+      cardinality: 'many',
+    },
+  ],
+  entries: [
+    {
+      sourceField: ['id'],
+      destField: ['accountId'],
+      destSchema: 'ledgerEntry',
+      cardinality: 'many',
+    },
+  ],
+  budgetLines: [
+    {
+      sourceField: ['id'],
+      destField: ['accountId'],
+      destSchema: 'budgetLine',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const ledgerEntryRelationships = {
+  transaction: [
+    {
+      sourceField: ['transactionId'],
+      destField: ['id'],
+      destSchema: 'ledgerTransaction',
+      cardinality: 'one',
+    },
+  ],
+  account: [
+    {
+      sourceField: ['accountId'],
+      destField: ['id'],
+      destSchema: 'ledgerAccount',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const ledgerTransactionRelationships = {
+  creator: [
+    {
+      sourceField: ['createdById'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
@@ -8342,73 +8164,133 @@ const timesheetRelationships = {
   entries: [
     {
       sourceField: ['id'],
-      destField: ['timesheetId'],
-      destSchema: 'timeEntry',
+      destField: ['transactionId'],
+      destSchema: 'ledgerEntry',
       cardinality: 'many',
     },
   ],
 } as const;
-const timeEntryRelationships = {
-  timesheet: [
+const marketingAudienceRelationships = {
+  campaignAudiences: [
     {
-      sourceField: ['timesheetId'],
-      destField: ['id'],
-      destSchema: 'timesheet',
-      cardinality: 'one',
-    },
-  ],
-  task: [
-    {
-      sourceField: ['taskId'],
-      destField: ['id'],
-      destSchema: 'projectTask',
-      cardinality: 'one',
+      sourceField: ['id'],
+      destField: ['audienceId'],
+      destSchema: 'marketingCampaignAudience',
+      cardinality: 'many',
     },
   ],
 } as const;
-const benefitPlanRelationships = {
-  administrator: [
+const marketingCampaignRelationships = {
+  owner: [
     {
-      sourceField: ['administratorId'],
+      sourceField: ['ownerId'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
     },
   ],
-  enrollments: [
+  channels: [
     {
       sourceField: ['id'],
-      destField: ['benefitPlanId'],
-      destSchema: 'benefitEnrollment',
+      destField: ['campaignId'],
+      destSchema: 'marketingCampaignChannel',
+      cardinality: 'many',
+    },
+  ],
+  audiences: [
+    {
+      sourceField: ['id'],
+      destField: ['campaignId'],
+      destSchema: 'marketingCampaignAudience',
       cardinality: 'many',
     },
   ],
 } as const;
-const benefitEnrollmentRelationships = {
-  benefitPlan: [
+const marketingCampaignAudienceRelationships = {
+  campaign: [
     {
-      sourceField: ['benefitPlanId'],
+      sourceField: ['campaignId'],
       destField: ['id'],
-      destSchema: 'benefitPlan',
+      destSchema: 'marketingCampaign',
       cardinality: 'one',
     },
   ],
-  employee: [
+  audience: [
     {
-      sourceField: ['employeeId'],
+      sourceField: ['audienceId'],
       destField: ['id'],
-      destSchema: 'employeeProfile',
+      destSchema: 'marketingAudience',
       cardinality: 'one',
     },
   ],
 } as const;
-const projectTagRelationships = {
-  taskLinks: [
+const marketingCampaignChannelRelationships = {
+  campaign: [
+    {
+      sourceField: ['campaignId'],
+      destField: ['id'],
+      destSchema: 'marketingCampaign',
+      cardinality: 'one',
+    },
+  ],
+  channel: [
+    {
+      sourceField: ['channelId'],
+      destField: ['id'],
+      destSchema: 'marketingChannel',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const marketingChannelRelationships = {
+  campaignChannels: [
     {
       sourceField: ['id'],
-      destField: ['tagId'],
-      destSchema: 'projectTaskTag',
+      destField: ['channelId'],
+      destSchema: 'marketingCampaignChannel',
       cardinality: 'many',
+    },
+  ],
+} as const;
+const mediumRelationships = {
+  messages: [
+    {
+      sourceField: ['id'],
+      destField: ['mediumId'],
+      destSchema: 'message',
+      cardinality: 'many',
+    },
+  ],
+  users: [
+    {
+      sourceField: ['id'],
+      destField: ['mediumId'],
+      destSchema: 'message',
+      cardinality: 'many',
+    },
+    {
+      sourceField: ['senderId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const messageRelationships = {
+  medium: [
+    {
+      sourceField: ['mediumId'],
+      destField: ['id'],
+      destSchema: 'medium',
+      cardinality: 'one',
+    },
+  ],
+  sender: [
+    {
+      sourceField: ['senderId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
     },
   ],
 } as const;
@@ -8454,6 +8336,88 @@ const projectRelationships = {
     },
   ],
 } as const;
+const projectAssignmentRelationships = {
+  task: [
+    {
+      sourceField: ['taskId'],
+      destField: ['id'],
+      destSchema: 'projectTask',
+      cardinality: 'one',
+    },
+  ],
+  user: [
+    {
+      sourceField: ['userId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const projectAttachmentRelationships = {
+  task: [
+    {
+      sourceField: ['taskId'],
+      destField: ['id'],
+      destSchema: 'projectTask',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const projectAuditRelationships = {
+  project: [
+    {
+      sourceField: ['projectId'],
+      destField: ['id'],
+      destSchema: 'project',
+      cardinality: 'one',
+    },
+  ],
+  actor: [
+    {
+      sourceField: ['actorId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const projectCommentRelationships = {
+  task: [
+    {
+      sourceField: ['taskId'],
+      destField: ['id'],
+      destSchema: 'projectTask',
+      cardinality: 'one',
+    },
+  ],
+  author: [
+    {
+      sourceField: ['authorId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const projectNoteRelationships = {
+  project: [
+    {
+      sourceField: ['projectId'],
+      destField: ['id'],
+      destSchema: 'project',
+      cardinality: 'one',
+    },
+  ],
+  author: [
+    {
+      sourceField: ['authorId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
 const projectPhaseRelationships = {
   project: [
     {
@@ -8468,6 +8432,16 @@ const projectPhaseRelationships = {
       sourceField: ['id'],
       destField: ['phaseId'],
       destSchema: 'projectTask',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const projectTagRelationships = {
+  taskLinks: [
+    {
+      sourceField: ['id'],
+      destField: ['tagId'],
+      destSchema: 'projectTaskTag',
       cardinality: 'many',
     },
   ],
@@ -8522,52 +8496,6 @@ const projectTaskRelationships = {
     },
   ],
 } as const;
-const projectAssignmentRelationships = {
-  task: [
-    {
-      sourceField: ['taskId'],
-      destField: ['id'],
-      destSchema: 'projectTask',
-      cardinality: 'one',
-    },
-  ],
-  user: [
-    {
-      sourceField: ['userId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const projectCommentRelationships = {
-  task: [
-    {
-      sourceField: ['taskId'],
-      destField: ['id'],
-      destSchema: 'projectTask',
-      cardinality: 'one',
-    },
-  ],
-  author: [
-    {
-      sourceField: ['authorId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const projectAttachmentRelationships = {
-  task: [
-    {
-      sourceField: ['taskId'],
-      destField: ['id'],
-      destSchema: 'projectTask',
-      cardinality: 'one',
-    },
-  ],
-} as const;
 const projectTaskTagRelationships = {
   task: [
     {
@@ -8582,42 +8510,6 @@ const projectTaskTagRelationships = {
       sourceField: ['tagId'],
       destField: ['id'],
       destSchema: 'projectTag',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const projectNoteRelationships = {
-  project: [
-    {
-      sourceField: ['projectId'],
-      destField: ['id'],
-      destSchema: 'project',
-      cardinality: 'one',
-    },
-  ],
-  author: [
-    {
-      sourceField: ['authorId'],
-      destField: ['id'],
-      destSchema: 'user',
-      cardinality: 'one',
-    },
-  ],
-} as const;
-const projectAuditRelationships = {
-  project: [
-    {
-      sourceField: ['projectId'],
-      destField: ['id'],
-      destSchema: 'project',
-      cardinality: 'one',
-    },
-  ],
-  actor: [
-    {
-      sourceField: ['actorId'],
-      destField: ['id'],
-      destSchema: 'user',
       cardinality: 'one',
     },
   ],
@@ -8672,6 +8564,42 @@ const supportTicketRelationships = {
     },
   ],
 } as const;
+const supportTicketAssignmentRelationships = {
+  ticket: [
+    {
+      sourceField: ['ticketId'],
+      destField: ['id'],
+      destSchema: 'supportTicket',
+      cardinality: 'one',
+    },
+  ],
+  assignee: [
+    {
+      sourceField: ['assigneeId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
+const supportTicketAuditRelationships = {
+  ticket: [
+    {
+      sourceField: ['ticketId'],
+      destField: ['id'],
+      destSchema: 'supportTicket',
+      cardinality: 'one',
+    },
+  ],
+  actor: [
+    {
+      sourceField: ['actorId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+} as const;
 const supportTicketMessageRelationships = {
   ticket: [
     {
@@ -8718,39 +8646,111 @@ const supportTicketTagLinkRelationships = {
     },
   ],
 } as const;
-const supportTicketAssignmentRelationships = {
-  ticket: [
+const teamRelationships = {
+  department: [
     {
-      sourceField: ['ticketId'],
+      sourceField: ['departmentId'],
       destField: ['id'],
-      destSchema: 'supportTicket',
+      destSchema: 'department',
       cardinality: 'one',
     },
   ],
-  assignee: [
+  lead: [
     {
-      sourceField: ['assigneeId'],
+      sourceField: ['leadId'],
       destField: ['id'],
       destSchema: 'user',
+      cardinality: 'one',
+    },
+  ],
+  employees: [
+    {
+      sourceField: ['id'],
+      destField: ['teamId'],
+      destSchema: 'employeeProfile',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const timeEntryRelationships = {
+  timesheet: [
+    {
+      sourceField: ['timesheetId'],
+      destField: ['id'],
+      destSchema: 'timesheet',
+      cardinality: 'one',
+    },
+  ],
+  task: [
+    {
+      sourceField: ['taskId'],
+      destField: ['id'],
+      destSchema: 'projectTask',
       cardinality: 'one',
     },
   ],
 } as const;
-const supportTicketAuditRelationships = {
-  ticket: [
+const timesheetRelationships = {
+  employee: [
     {
-      sourceField: ['ticketId'],
+      sourceField: ['employeeId'],
       destField: ['id'],
-      destSchema: 'supportTicket',
+      destSchema: 'employeeProfile',
       cardinality: 'one',
     },
   ],
-  actor: [
+  submittedBy: [
     {
-      sourceField: ['actorId'],
+      sourceField: ['submittedById'],
       destField: ['id'],
       destSchema: 'user',
       cardinality: 'one',
+    },
+  ],
+  entries: [
+    {
+      sourceField: ['id'],
+      destField: ['timesheetId'],
+      destSchema: 'timeEntry',
+      cardinality: 'many',
+    },
+  ],
+} as const;
+const userRelationships = {
+  messages: [
+    {
+      sourceField: ['id'],
+      destField: ['senderId'],
+      destSchema: 'message',
+      cardinality: 'many',
+    },
+  ],
+  mediums: [
+    {
+      sourceField: ['id'],
+      destField: ['senderId'],
+      destSchema: 'message',
+      cardinality: 'many',
+    },
+    {
+      sourceField: ['mediumId'],
+      destField: ['id'],
+      destSchema: 'medium',
+      cardinality: 'many',
+    },
+  ],
+  friends: [
+    {
+      sourceField: ['id'],
+      destField: ['requestingId'],
+      destSchema: 'friendship',
+      cardinality: 'many',
+    },
+    {
+      sourceField: ['acceptingId'],
+      destField: ['id'],
+      destSchema: 'user',
+      cardinality: 'many',
     },
   ],
 } as const;
@@ -8864,65 +8864,65 @@ export const schema = {
     orderPayment: orderPaymentRelationships,
     shipment: shipmentRelationships,
     shipmentItem: shipmentItemRelationships,
-    user: userRelationships,
-    medium: mediumRelationships,
-    message: messageRelationships,
-    filters: filtersRelationships,
-    crmAccount: crmAccountRelationships,
-    crmContact: crmContactRelationships,
-    crmPipelineStage: crmPipelineStageRelationships,
-    crmOpportunity: crmOpportunityRelationships,
-    crmOpportunityStageHistory: crmOpportunityStageHistoryRelationships,
-    crmActivityType: crmActivityTypeRelationships,
-    crmActivity: crmActivityRelationships,
-    crmNote: crmNoteRelationships,
-    documentLibrary: documentLibraryRelationships,
-    documentFolder: documentFolderRelationships,
-    documentFile: documentFileRelationships,
-    documentFileVersion: documentFileVersionRelationships,
-    documentSharing: documentSharingRelationships,
+    benefitEnrollment: benefitEnrollmentRelationships,
+    benefitPlan: benefitPlanRelationships,
     billingInvoice: billingInvoiceRelationships,
     billingInvoiceLine: billingInvoiceLineRelationships,
-    expenseReport: expenseReportRelationships,
-    expenseItem: expenseItemRelationships,
-    ledgerAccount: ledgerAccountRelationships,
-    ledgerTransaction: ledgerTransactionRelationships,
-    ledgerEntry: ledgerEntryRelationships,
     budget: budgetRelationships,
     budgetLine: budgetLineRelationships,
-    integrationWebhook: integrationWebhookRelationships,
-    integrationEvent: integrationEventRelationships,
-    integrationCredential: integrationCredentialRelationships,
-    marketingCampaign: marketingCampaignRelationships,
-    marketingChannel: marketingChannelRelationships,
-    marketingCampaignChannel: marketingCampaignChannelRelationships,
-    marketingAudience: marketingAudienceRelationships,
-    marketingCampaignAudience: marketingCampaignAudienceRelationships,
+    crmAccount: crmAccountRelationships,
+    crmActivity: crmActivityRelationships,
+    crmActivityType: crmActivityTypeRelationships,
+    crmContact: crmContactRelationships,
+    crmNote: crmNoteRelationships,
+    crmOpportunity: crmOpportunityRelationships,
+    crmOpportunityStageHistory: crmOpportunityStageHistoryRelationships,
+    crmPipelineStage: crmPipelineStageRelationships,
     department: departmentRelationships,
-    employeeProfile: employeeProfileRelationships,
-    team: teamRelationships,
-    employmentHistory: employmentHistoryRelationships,
+    documentFile: documentFileRelationships,
+    documentFileVersion: documentFileVersionRelationships,
+    documentFolder: documentFolderRelationships,
+    documentLibrary: documentLibraryRelationships,
+    documentSharing: documentSharingRelationships,
     employeeDocument: employeeDocumentRelationships,
-    timesheet: timesheetRelationships,
-    timeEntry: timeEntryRelationships,
-    benefitPlan: benefitPlanRelationships,
-    benefitEnrollment: benefitEnrollmentRelationships,
-    projectTag: projectTagRelationships,
+    employeeProfile: employeeProfileRelationships,
+    employmentHistory: employmentHistoryRelationships,
+    expenseItem: expenseItemRelationships,
+    expenseReport: expenseReportRelationships,
+    filters: filtersRelationships,
+    integrationCredential: integrationCredentialRelationships,
+    integrationEvent: integrationEventRelationships,
+    integrationWebhook: integrationWebhookRelationships,
+    ledgerAccount: ledgerAccountRelationships,
+    ledgerEntry: ledgerEntryRelationships,
+    ledgerTransaction: ledgerTransactionRelationships,
+    marketingAudience: marketingAudienceRelationships,
+    marketingCampaign: marketingCampaignRelationships,
+    marketingCampaignAudience: marketingCampaignAudienceRelationships,
+    marketingCampaignChannel: marketingCampaignChannelRelationships,
+    marketingChannel: marketingChannelRelationships,
+    medium: mediumRelationships,
+    message: messageRelationships,
     project: projectRelationships,
-    projectPhase: projectPhaseRelationships,
-    projectTask: projectTaskRelationships,
     projectAssignment: projectAssignmentRelationships,
-    projectComment: projectCommentRelationships,
     projectAttachment: projectAttachmentRelationships,
-    projectTaskTag: projectTaskTagRelationships,
-    projectNote: projectNoteRelationships,
     projectAudit: projectAuditRelationships,
+    projectComment: projectCommentRelationships,
+    projectNote: projectNoteRelationships,
+    projectPhase: projectPhaseRelationships,
+    projectTag: projectTagRelationships,
+    projectTask: projectTaskRelationships,
+    projectTaskTag: projectTaskTagRelationships,
     supportTicket: supportTicketRelationships,
+    supportTicketAssignment: supportTicketAssignmentRelationships,
+    supportTicketAudit: supportTicketAuditRelationships,
     supportTicketMessage: supportTicketMessageRelationships,
     supportTicketTag: supportTicketTagRelationships,
     supportTicketTagLink: supportTicketTagLinkRelationships,
-    supportTicketAssignment: supportTicketAssignmentRelationships,
-    supportTicketAudit: supportTicketAuditRelationships,
+    team: teamRelationships,
+    timeEntry: timeEntryRelationships,
+    timesheet: timesheetRelationships,
+    user: userRelationships,
   },
   enableLegacyQueries: false,
   enableLegacyMutators: false,
